% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverk.R
\name{add_fig}
\alias{add_fig}
\title{A function that adds a plotly fig to a datapackage object.}
\usage{
add_fig(dp, fig, title, description = "")
}
\arguments{
\item{dp}{Datapackage}

\item{fig}{Plotly figure}

\item{title}{Title of figure}

\item{description}{Description of figure}
}
\description{
A function that adds a plotly fig to a datapackage object.
}
