% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverk.R
\name{add_resource}
\alias{add_resource}
\title{A function that adds a dataframe resource to a datapackage object.}
\usage{
add_resource(
  dp,
  dataframe,
  resource_name,
  resource_description = "",
  python_path = system("which python")
)
}
\arguments{
\item{dp}{Datapackage}

\item{resource_name}{Name of the resource}

\item{resource_description}{Description of the dataframe}

\item{python_path}{Path to python version}

\item{resource}{Dataframe to be added}
}
\description{
A function that adds a dataframe resource to a datapackage object.
}
