% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverk.R
\name{create_dp}
\alias{create_dp}
\title{A function that creates a dataverk datapackage object.}
\usage{
create_dp(metadata, python_path = system("which python"))
}
\arguments{
\item{metadata}{Metadata}

\item{python_path}{Path to python version}
}
\description{
A function that creates a dataverk datapackage object.
}
