% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverk.R
\name{publish}
\alias{publish}
\title{A function that publishes a datapackage.}
\usage{
publish(dp, python_path = system("which python"))
}
\arguments{
\item{dp}{Datapackage to be published}

\item{python_path}{Path to python version}
}
\description{
A function that publishes a datapackage.
}
