% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fig.R
\name{add_fig}
\alias{add_fig}
\title{Adds plotly fig to datapackage object}
\usage{
add_fig(dp, fig, title = NULL, description = NULL)
}
\arguments{
\item{dp}{Datapackage}

\item{fig}{Plotly figure}

\item{title}{Title of figure}

\item{description}{Description of figure}
}
\description{
Adds plotly fig to datapackage object
}
\examples{
# Importing libraries
library(dataverkr)
library(plotly)

# Creating dataframe
country <- c("Norge","Sverige", "Danmark")
numberOf <- c(200, 300, 500)
df <- data.frame(country, numberOf)

# Creating plotly figure
fig <-  plot_ly(df, x = ~country, y = ~numberOf, type = 'bar')
fig


# Metadata for datapackage
metadata <- list(title = 'DataverkR example',
                 bucket =  'nav-opendata',
                 store = 'nais',
                 temporal = list(from = '2020-09-16',
                                 to='2020-09-16'),
                 type =  'datapackage',
                 format ='datapackage')

# Creating datapackage
dp <- dataverkr::create_dp(metadata)

# Adding plotly figure to the datapackage
dataverkr::add_fig(dp = dp,
                   fig = fig,
                   title = 'Some plotly figure',
                   description = 'This is a plotly figure example')

# Adding dataframe as a resource to the datapackage
dataverkr::add_resource(dp = dp,
                        dataframe = df,
                        resource_name = 'some_dataframe',
                        resource_description = 'This is a resource example')

# Publishing the datapackage
dataverkr::publish(dp)
}
